llmpoc/requirements.txt
numpy
pyarrow
pandas
transformers
accelerate
openpyxl
djangorestframework
django-cors-headers